Rename `bitmappool` package to `bitmap`.
