+++ coil-base/src/androidTest/java/coil/RealImageLoaderTest.kt
@@ -15,7 +15,7 @@ import androidx.core.net.toUri
 import androidx.test.core.app.ApplicationProvider
 import coil.annotation.ExperimentalCoilApi
 import coil.base.test.R
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.decode.BitmapFactoryDecoder
 import coil.decode.DecodeResult
 import coil.decode.Decoder
