+++ coil-base/src/androidTest/java/coil/transform/CircleCropTransformationTest.kt
@@ -2,7 +2,7 @@ package coil.transform
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.size.OriginalSize
 import coil.util.decodeBitmapAsset
 import coil.util.isSimilarTo
