+++ coil-base/src/androidTest/java/coil/transform/GrayscaleTransformationTest.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import androidx.test.core.app.ApplicationProvider
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.size.OriginalSize
 import coil.util.decodeBitmapAsset
 import coil.util.getPixels
