+++ coil-base/src/main/java/coil/decode/EmptyDecoder.kt
@@ -2,7 +2,7 @@ package coil.decode
 
 import android.graphics.drawable.ColorDrawable
 import coil.ComponentRegistry
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.size.Size
 import okio.BufferedSource
 import okio.blackholeSink
