+++ coil-base/src/main/java/coil/fetch/DrawableFetcher.kt
@@ -2,7 +2,7 @@ package coil.fetch
 
 import android.content.Context
 import android.graphics.drawable.Drawable
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.decode.DataSource
 import coil.decode.DrawableDecoderService
 import coil.decode.Options
