+++ coil-base/src/main/java/coil/fetch/HttpUrlFetcher.kt
@@ -1,7 +1,7 @@
 package coil.fetch
 
 import android.webkit.MimeTypeMap
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.decode.DataSource
 import coil.decode.Options
 import coil.network.HttpException
