+++ coil-base/src/test/java/coil/bitmap/SizeStrategyTest.kt
@@ -1,4 +1,4 @@
-package coil.bitmappool
+package coil.bitmap
 
 import android.graphics.Bitmap
 import coil.util.createBitmap
@@ -9,13 +9,13 @@ import org.robolectric.RobolectricTestRunner
 import kotlin.test.assertEquals
 
 @RunWith(RobolectricTestRunner::class)
-class AttributeStrategyTest {
+class SizeStrategyTest {
 
-    private lateinit var strategy: AttributeStrategy
+    private lateinit var strategy: SizeStrategy
 
     @Before
     fun before() {
-        strategy = AttributeStrategy()
+        strategy = SizeStrategy()
     }
 
     @Test
@@ -35,18 +35,20 @@ class AttributeStrategyTest {
     }
 
     @Test
-    fun `large enough bitmap with different config is not reused`() {
+    fun `large enough bitmap with different config is reused`() {
         val bitmap = createBitmap(width = 250, height = 250, config = Bitmap.Config.RGB_565)
         strategy.put(bitmap)
 
-        assertEquals(null, strategy.get(100, 100, Bitmap.Config.ARGB_8888))
+        assertEquals(bitmap, strategy.get(100, 100, Bitmap.Config.ARGB_8888))
     }
 
     @Test
-    fun `large enough bitmap with different size is not reused`() {
-        val bitmap = createBitmap(width = 150, height = 150)
-        strategy.put(bitmap)
+    fun `only puts are factored into ceilingKey`() {
+        strategy.put(createBitmap(100, 100, Bitmap.Config.ARGB_8888))
+        strategy.get(200, 100, Bitmap.Config.ARGB_8888)
+        val expected = createBitmap(300, 100, Bitmap.Config.ARGB_8888)
+        strategy.put(expected)
 
-        assertEquals(null, strategy.get(100, 100, Bitmap.Config.ARGB_8888))
+        assertEquals(expected, strategy.get(200, 100, Bitmap.Config.ARGB_8888))
     }
 }
