+++ coil-base/src/test/java/coil/fetch/HttpUrlFetcherTest.kt
@@ -2,7 +2,7 @@ package coil.fetch
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
-import coil.bitmappool.BitmapPool
+import coil.bitmap.BitmapPool
 import coil.size.PixelSize
 import coil.util.createMockWebServer
 import coil.util.createOptions
